Fix invalid schema when type has no required properties.
The Open API / Swagger specification states that if there
are no required properties, you must not include the required
property. It explicitly states that an empty array is invalid,
and the online editor also complains if required is null.

Change to remove the required property when initializing the
details object, it will only be added if a required property
exists on the type.